-- project structure reorg -- bugfix -- gwt2 and j2cl tests added -- ready for gwt3
