+++ .gitignore
@@ -1,0 +1,6 @@
+*.class
+.project
+.settings/
+.classpath
+.idea/
+*.iml
