+++ README.md
@@ -1,0 +1,30 @@
+# gwt-xml
+GWT XMLHttpRequest module
+
+
+![GWT3/J2CL compatible](https://img.shields.io/badge/GWT3/J2CL-compatible-brightgreen.svg)
+
+# GWT XML
+
+### Dependency
+
+```xml
+<dependency>
+    <groupId>org.gwtproject.xml</groupId>
+    <artifactId>gwt-xml</artifactId>
+    <version>1.0-SNAPSHOT</version>
+</dependency>
+```
+
+### Instructions
+To build gwt-xml:
+
+* run `mvn clean install`
+
+on the parent directory.
+
+To run the j2cl tests:
+
+* switch to the 'gwt-xml-j2cl-tests' directory
+* run `mvn j2cl:clean` & `mvn j2cl:test`
+
