+++ gwt-xml-gwt2-tests/src/test/resources/org/gwtproject/xml/XMLTest.gwt.xml
@@ -1,0 +1,27 @@
+<!--                                                                        -->
+<!-- Copyright 2009 Google Inc.                                             -->
+<!-- Licensed under the Apache License, Version 2.0 (the "License"); you    -->
+<!-- may not use this file except in compliance with the License. You may   -->
+<!-- may obtain a copy of the License at                                    -->
+<!--                                                                        -->
+<!-- http://www.apache.org/licenses/LICENSE-2.0                             -->
+<!--                                                                        -->
+<!-- Unless required by applicable law or agreed to in writing, software    -->
+<!-- distributed under the License is distributed on an "AS IS" BASIS,      -->
+<!-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or        -->
+<!-- implied. License for the specific language governing permissions and   -->
+<!-- limitations under the License.                                         -->
+
+<!-- Types and resources required to support HTTP requests.                 -->
+<!--                                                                        -->
+<!-- Any user application that wishes to use the types declared in the http -->
+<!-- package must inherit this module.                                      -->
+<!--                                                                        -->
+<module>
+
+  <!-- Other module inherits                                      -->
+  <inherits name="org.gwtproject.xml.XML"/>
+
+  <!-- sources -->
+  <source path="client"/>
+</module>
