+++ gwt-xml/src/main/java/org/gwtproject/xml/COPYING
