+++ gwt-xml/src/main/java/org/gwtproject/xml/client/Comment.java
