+++ gwt-xml/src/main/java/org/gwtproject/xml/client/DOMException.java
