+++ gwt-xml/src/main/java/org/gwtproject/xml/client/XMLParser.java
