+++ gwt-xml/src/main/java/org/gwtproject/xml/client/impl/ElementImpl.java
