+++ gwt-xml/src/main/resources/org/gwtproject/xml/XML.gwt.xml
@@ -16,5 +16,6 @@
 <!--                                                                        -->
 <module>
   <inherits name="elemental2.dom.Dom"/>
+  <inherits name="jsinterop.base.Base"/>
   <source path="client" />
 </module>
